test(install): consolidate OEM FastOS regression
